Moved phy.waveform to phy.traces.
